requirements.txt
fastapi
uvicorn
langchain-upstage
python-dotenv
